drawnet/requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
Pillow
tqdm
pyyaml
jupyter
kaggle
requests
grad-cam
